fix: use the short url to connections notify
